+++ R/googleAuthR_auth.R
@@ -14,26 +14,18 @@ Authentication <- R6::R6Class(
   parent_env = emptyenv()
 )
 
-## Set scopes via:
-# options("googleAuthR.scopes.selected" = c("https://www.googleapis.com/auth/webmasters",
-#                                           "https://www.googleapis.com/auth/analytics",
-#                                           "https://www.googleapis.com/auth/analytics.readonly",
-#                                           "https://www.googleapis.com/auth/analytics.manage.users.readonly",
-#                                           "https://www.googleapis.com/auth/tagmanager.readonly"))
-
 #' Authorize \code{googleAuthR}
 #'
 #' Authorize \code{googleAuthR} to access your Google user data. You will be
 #' directed to a web browser, asked to sign in to your Google account, and to
-#' grant \code{googleAuthR} access to user data for Google Search Console. These user credentials are cached in a file named
+#' grant \code{googleAuthR} access to user data for Google Search Console. 
+#' These user credentials are cached in a file named
 #' \code{.httr-oauth} in the current working directory, from where they can be
 #' automatically refreshed, as necessary.
 #'
 #'
-#' In a call to \code{gar_auth}, the user can provide the token, app key and
-#' secret explicitly and can dictate whether credentials will be cached in
-#' \code{.httr_oauth}. If unspecified, these arguments are controlled via
-#' options, which, if undefined at the time \code{googleAuthR} is loaded, are
+#' These arguments are controlled via options, which, 
+#' if undefined at the time \code{googleAuthR} is loaded, are
 #' defined like so:
 #'
 #' \describe{